Revert all build-related changes
Causing confusion and not comprehensively tested before being pushed. Additionally, makes things harder to work with for no clear reason

Please don't make major workflow changes in the middle of the night when nobody can check that they even work, particularly if they're being done just for the sake of it.