Fix XP start button text (hopefully)
This is a better hack to prevent the XP start button text from clipping.
It's a bit complicated and required hooking comctl32 v6 in order to
achieve.

comctl32!Button_DrawThemed (in v6) is responsible for rendering themed
buttons. Windows 10 seems to have some extra margin code that calls into
GetTextExtentPoint32W.

The gist of the hook is as follows:
- We find the HWND of the start button in uxtheme!SetWindowTheme, which
  provides a reliable method to search for the start button using its
  unique theme class.
- We hook comctl32!Button_DrawThemed in order to compare the HWND of the
  start button while it draws. You cannot hook DrawThemeText for a similar
  effect, as it passes a memory DC without a HWND. If the button's HWND is
  equal to that of the start button, then we enable the
  GetTextExtentPoint32W hook.
- We also hook gdi32!GetTextExtentPoint32W in order to coerce the drawing
  bounds of the start button text to a wider amount. To prevent any adverse
  effects elsewhere, this hook is only enabled if the caller by
  Button_DrawThemed when the HWND is the start button.

Some patterns may need to be adjusted for use across multiple operating
systems. I have only tested on Windows 10.0.19042.

Signed-off-by: Isabella <kawapure@gmail.com>