Fix explorer startup lag + system metrics changes
XP explorer used to take an unholy amount of time to start up (~30 seconds
on a heavy-use system), brought down to just 1 to bring the taskbar on
screen.

The lag was because of some functions called early in Explorer's
initialisation: CheckDefaultUIFonts & ChangeUIfontsToNewDPI.

These both import and call desk.cpl functions which mess with system
metrics. They seem to only exist for OOBE adjustments and automatically
adjusting to hardware configuration changes. On Windows 10, it seems that
keeping this behaviour around only causes adverse effects, so it's better
to just patch it away.

Signed-off-by: Isabella <kawapure@gmail.com>