Fix grouped app menus
Grouped app menus would previously immediately close upon any mouse click.
This appears to be a bug introduced in shell32 in some version of Windows
after XP. The supposed bug is related to some message filtering hook
function installed for menu bands, for which I don't know the purpose.

I opted to disable the functionality altogether, which seems to fix the
problems with opening the control with seemingly no adverse side effects.

By the way, you could have always used the menus using keyboard navigation,
as it was not affected by this issue.

Signed-off-by: Isabella <kawapure@gmail.com>