Improve CTray::SyncThreadProc pattern
The previous one had hardcoded pointers, so it only worked on the en-US
binary. Now it should work on multiple languages.

This was an important pattern to modify since the hook function is used to
retrieve a thread ID for bounds checking across the program. If this hook
were to fail, then we'll have incorrect behaviour across the whole program.

Signed-off-by: Isabella <kawapure@gmail.com>