Fix ClassicTheme regkey for hopefully the last time
The DWM frame will no longer appear if the user is using Classic or Basic themes.

Why it works like this and not through applying changes to DwmSetWindowAttribute is a mystery to me. Anyway...

Additionally restore commented out code for classic start icon that was inadvertently removed during the HookAPIs cleanup some days ago.