BUGFIX: Ensure pseudo-aero fallback on pre-1803 where there is no Win32 acrylic effect yet
Additionally, added a couple of extra comments. Fallback added to ensure the user does not encounter the "full #00000000 colorization fallback" which is not good UX.

The reason we fall back to Pseudoaero on 1709 and earlier in place of Acrylic, rather than BlurBehind is because of 2 reasons:
1. More OS consistency (OS taskbar was translucent from 8.0-1709)
2. BlurBehind tends to be buggy with taskbar behaviour (Supposed API limitation)

As for why we use Acrylic in-place of BlurBehind on Nickel and higher, it is less prone to issues than the latter, and also ensures greater OS consistency as a fallback for Win11.

So hopefully that explains the two varying approaches to fallback behaviour that we use here.

I'm currently considering shifting the enumeration a little so that it goes Legacy->Pseudoaero->Solid->BlurBehind->Acrylic but I might wait until we remove current mode 0 in Milestone 3 before doing that so that only 2 of the modes switch places rather than nearly all of them.

Plus, we need to focus on more bug-fixes like this one than messing with things. Only exception might be finishing up RPEnabled orb in future. Anyway I'm rambling on so I should stop typing.

~Grace