Add option to override colour alpha
Due to Windows 10+ colorization limitations, added option to allow user to change colour opacity without affecting system settings or having the chance of the OS overriding user preference.

There are two parts to this:
- The value to enable the override (defaulted to false)
- The hex value for the override as a DWORD (if there isn't one, falls back to Win7 default alpha value).

Wasn't going to add this but given the Windows tendency to reset the alpha value in the DWM ColorizationColor value, I wanted to ensure that user choice remained available.

~Grace