ShellHook implementation for Immersive Shell hotkeys
This is the third attempt at doing this. This time however, making use of my latest method to block the initial menu appearance. Integration is included with all supported Windows 10 and 11 versions. This is not yet completed, subject to some final testing on early versions of OS build 22621.

This resolves the long-standing issue with the immersive shell and the start menu hotkey.

Some compiler warnings have also been disabled from here on out as these specifically relate to things that don't negatively affect the program's functionality.

Some left-over code remains from earlier attempts at fixing the issue that will need to be pruned in the near future.

~Grace