Resolve issues with the search UI in the latest Windows 11 updates
In recent builds of Windows 11, the ItemsView interface used for File Explorer (and for the start menu search) was updated, with the DirectUI markup now mandating a completely new class known as `ItemsViewAccessible::Header`.

This caused problems for us, as the start menu search grinds to a halt because the class cannot be read from the inactive style.

Given limitations that remain on the interoperability of modern system-wide visual styles and the explorer7 start menu, moving away from this model was not an option.

After spending some time thinking over it, the first approach that made sense to check was forcing the system style to be used for the search interface. I'd discovered that search uses a different theming API call to other components, and because I'd investigated undocumented APIs for the Windows visual style system in the past, I knew how to check for the class definition. As such, something like this was plausible, and *did* in fact work. However, very quickly it became apparent that functionality would either have to change for all users, or for Windows 11 users specifically, in such a way that would lessen consistency with the user-provided theme.

Ultimately, with this in mind, I started again and had something of a re-think, in deciding to instead investigate the way the DirectUI themeable calls are handled, as Win32 APIs are used under-the-hood. Research then showed that the standard theme APIs were used, and after digging through things further, I decided to test what would happen if we simply only allowed the inactive style loader to be used (within the applicable API call) when the class specified was not `ItemsViewAccessible::Header` - on the surface, this finally restored the desired functionality.

So then, Microsoft... all that for a drop of blood. Was it worth it?

~Grace