Revise ShellHook procedure + light clean-ups
The change here is to shift towards wiring it to the message for opening the Windows 7 start menu directly - this is because CTaskBand::HandleShellHook is unreliable for some versions, and we should avoid unnecessary build checks where applicable.

XamlLauncher::OnShellHookMessage does itself fire straight to XamlLauncher::ShowStartView in modern Windows (when applicable), so the logic is sound. As per testing, seemingly no issues with this.

Some light cleanups have also been applied. More will be coming imminently.

~Grace