Fix a crashing issue with IUnknown_QueryServiceExecNEW
Oddly this issue was only introduced by the type corrections done to a series of components relating to the start menu, but thankfully the fix itself was fairly simple and only involved having to add an additional check for a parameter.

The bug itself for reference: clicking an empty section of the start menu would cause explorer to crash.

This commit also corrects the syntax of one GUID name, increments the version and adds the pre-release registry configuration code differentiator.