Fix immersive settings wrongly getting identified as Win32
In essence, this AppID filter ensures that the immersive settings application does not wrongly get assigned the dud shortcut that bears no relation to it other than using a similar-looking icon.

As such, it should now behave in line with other immersive applications, which remains fundamentally flawed at this stage but at the very least should help to reduce confusion and improve consistency.

A more substantial change will be shipped in Milestone 3