Allow pinned immersive applications to be hidden
This expands the scope of the StoreAppsOnTaskbar registry value to now hide pinned immersive applications, on top of the existing functionality of disabling the icon loader for unpinned immersive applications.

This also removes some dead code for attempting to change the pinnable information for immersive applications (resolved better in Milestone 3) and unused invocation of PinManager interface (implemented properly in Milestone 3).