Update IsForbiddenShortcut for greater consistency with TH1/TH2
MiracastView and "Print Dialog" are also filtered out by some means in the original TH1/TH2 program list that ours is based off of, therefore they are now also filtered.

Moreover, the forbidden shortcut filter is now only applied to the immersive-inclusive version of the program list, as no filter is applied on Windows 8 or 8.1's earlier iteration of the program list.

A lot of words for something fairly simple. Nevertheless...