Add IsForbiddenShortcut to the start menu item filter
The fake "Settings" (actually Control Panel), MiracastView and "Print Dialog" shortcuts are filtered out by some means in the original TH1/TH2 program list that ours is based off of, therefore they are now filtered. This should improve the user experience as these are not ordinarily visible in the modern start menu, and may have caused users some confusion.

Moreover, the forbidden shortcut filter is only applied to the immersive-inclusive version of the program list, as no filter is applied on Windows 8 or 8.1's earlier iteration of the program list.